+++ lib/zebra_emdk_plugin.dart
@@ -2,6 +2,7 @@ export 'services/emdk_manager.dart';
 export 'services/barcode_manager.dart';
 export 'services/notification_manager.dart';
 export 'services/profile_manager.dart';
+export 'services/key_event_manager.dart';
 
 export 'custom_models/connection_change.dart';
 
@@ -11,3 +12,4 @@ export 'generated/scan_data.dart';
 export 'generated/scanner_config.dart';
 export 'generated/scanner.dart';
 export 'generated/decoder_params.dart';
+export 'generated/key_identifiers.dart';